+++ src/core/session/index.ts
@@ -7,42 +7,43 @@
  */
 
 import session from 'core/session/engines';
+import { emitter } from 'core/session/const';
 import { Session, SessionKey } from 'core/session/interface';
-import { EventEmitter2 as EventEmitter } from 'eventemitter2';
 export * from 'core/session/interface';
 
-export const
-	event = new EventEmitter({maxListeners: 100, newListener: false});
-
 /**
- * Returns current session object
+ * Returns information about the current session
  */
 export async function get(): Promise<Session> {
 	try {
 		const
 			s = await session;
 
+		const [auth, params] = await Promise.all([
+			s.get<SessionKey>('auth'),
+			s.get<Dictionary>('params')
+		]);
+
 		return {
-			auth: await s.get<SessionKey>('auth'),
-			csrf: await s.get<SessionKey>('csrf')
+			auth,
+			params
 		};
 
 	} catch {
 		return {
-			auth: undefined,
-			csrf: undefined
+			auth: undefined
 		};
 	}
 }
 
 /**
- * Sets a new session
+ * Sets a new session with the specified parameters
  *
  * @param [auth]
- * @param [csrf]
+ * @param [params] - additional parameters
  * @emits set(session: Session)
  */
-export async function set(auth?: SessionKey, csrf?: SessionKey): Promise<boolean> {
+export async function set(auth?: SessionKey, params?: Dictionary): Promise<boolean> {
 	try {
 		const
 			s = await session;
@@ -51,11 +52,11 @@ export async function set(auth?: SessionKey, csrf?: SessionKey): Promise<boolean
 			await s.set('auth', auth);
 		}
 
-		if (csrf) {
-			await s.set('csrf', csrf);
+		if (params) {
+			await s.set('params', params);
 		}
 
-		event.emit('set', {auth, csrf});
+		emitter.emit('set', {auth, params});
 
 	} catch {
 		return false;
@@ -65,15 +66,14 @@ export async function set(auth?: SessionKey, csrf?: SessionKey): Promise<boolean
 }
 
 /**
- * Clears current session
+ * Clears the current session
  * @emits clear()
  */
 export async function clear(): Promise<boolean> {
 	try {
 		const s = await session;
-		await s.remove('auth');
-		await s.remove('csrf');
-		event.emit('clear');
+		await Promise.all([s.remove('auth'), s.remove('params')]);
+		emitter.emit('clear');
 
 	} catch {
 		return false;
@@ -83,15 +83,15 @@ export async function clear(): Promise<boolean> {
 }
 
 /**
- * Matches the specified session and current
+ * Matches a session with the current
  *
  * @param [auth]
- * @param [csrf]
+ * @param [params]
  */
-export async function match(auth?: SessionKey, csrf?: SessionKey): Promise<boolean> {
+export async function match(auth?: SessionKey, params?: Dictionary): Promise<boolean> {
 	try {
 		const s = await get();
-		return auth === s.auth && csrf === s.csrf;
+		return auth === s.auth && (params === undefined || Object.fastCompare(params, s.params));
 
 	} catch {
 		return false;
@@ -99,12 +99,11 @@ export async function match(auth?: SessionKey, csrf?: SessionKey): Promise<boole
 }
 
 /**
- * Returns true if the session object is exists
+ * Returns true if a session is already exists
  */
 export async function isExists(): Promise<boolean> {
 	try {
-		const s = await get();
-		return Boolean(s.auth && (!await (await session).has('csrf') || s.csrf));
+		return Boolean((await get()).auth);
 
 	} catch {
 		return false;