+++ CHANGELOG.md
@@ -27,6 +27,8 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 * Removed `core/thread`
 * Promise versions of the `Async` methods now returns a SyncPromise instance
 * Removed hardcode for `document.addEventListener` options within `Async` methods
+* Removed the legacy CSRF header from the session API
+* Added a new header the session API: it's can takes a dictionary with additional parameters  
 
 #### :rocket: New Feature